+++ java/vector/src/main/java/org/apache/arrow/vector/util/FieldVectorSerdeUtilitiy.java
@@ -20,6 +20,7 @@ package org.apache.arrow.vector.util;
 import io.netty.buffer.ArrowBuf;
 import org.apache.arrow.vector.BaseFixedWidthVector;
 import org.apache.arrow.vector.BaseVariableWidthVector;
+import org.apache.arrow.vector.BitVectorHelper;
 import org.apache.arrow.vector.FieldVector;
 import org.apache.arrow.vector.ipc.ReadChannel;
 import org.apache.arrow.vector.ipc.WriteChannel;
@@ -28,28 +29,44 @@ import org.apache.arrow.vector.ipc.message.MessageSerializer;
 import java.io.IOException;
 
 /**
- * Utility class for FieldVector.
+ * Utility class for serializing/deserializing FieldVector(only support BaseFixedWidthVector and BaseVariableWidthVector).
+ * The serialization format is:
+ * <pre>
+ * | 4 byte for valueCount
+ * | 4 byte for ArrowBuf capacity | serialized first ArrowBuf
+ * ...
+ * | 4 byte for ArrowBuf capacity | serialized nth ArrowBuf
+ * </pre>
+ * Note: This isn't part of the arrow standard format.
  */
-public class FieldVectorUtility {
+public class FieldVectorSerdeUtilitiy {
 
+  private FieldVectorSerdeUtilitiy(){}
+
+  /**
+   * write the vector to a channel.
+   * @param out the channel to write data
+   * @param vector the vector to write
+   * @throws IOException
+   */
   public static void serializeFieldVector(WriteChannel out, FieldVector vector) throws IOException {
     int valueCount = vector.getValueCount();
-    byte[] reuseIntBytes = new byte[4];
-    MessageSerializer.intToBytes(valueCount, reuseIntBytes);
-    out.write(reuseIntBytes);
+    byte[] sizeBytes = new byte[4];
+    MessageSerializer.intToBytes(valueCount, sizeBytes);
+    out.write(sizeBytes);
     ArrowBuf dataBuffer = vector.getDataBuffer();
     ArrowBuf validityBuffer = vector.getValidityBuffer();
 
-    int validityBufferCapacity = (valueCount + 7) >> 3;
+    int validityBufferCapacity = BitVectorHelper.getValidityBufferSize(valueCount);
 
     if (vector instanceof BaseFixedWidthVector) {
 
       int dataBufferCapacity = vector.getBufferSizeFor(valueCount) - validityBufferCapacity;
 
-      MessageSerializer.intToBytes(dataBufferCapacity, reuseIntBytes);
-      out.write(reuseIntBytes);
-      MessageSerializer.intToBytes(validityBufferCapacity, reuseIntBytes);
-      out.write(reuseIntBytes);
+      MessageSerializer.intToBytes(dataBufferCapacity, sizeBytes);
+      out.write(sizeBytes);
+      MessageSerializer.intToBytes(validityBufferCapacity, sizeBytes);
+      out.write(sizeBytes);
 
       out.write(dataBuffer.slice(0, dataBufferCapacity));
       out.write(validityBuffer.slice(0, validityBufferCapacity));
@@ -59,34 +76,39 @@ public class FieldVectorUtility {
       int offsetBufferCapacity = (valueCount + 1) * BaseVariableWidthVector.OFFSET_WIDTH;
       int dataBufferCapacity = offsetBuffer.getInt(valueCount * 4);
 
-      MessageSerializer.intToBytes(dataBufferCapacity, reuseIntBytes);
-      out.write(reuseIntBytes);
-      MessageSerializer.intToBytes(validityBufferCapacity, reuseIntBytes);
-      out.write(reuseIntBytes);
-      MessageSerializer.intToBytes(offsetBufferCapacity, reuseIntBytes);
-      out.write(reuseIntBytes);
+      MessageSerializer.intToBytes(dataBufferCapacity, sizeBytes);
+      out.write(sizeBytes);
+      MessageSerializer.intToBytes(validityBufferCapacity, sizeBytes);
+      out.write(sizeBytes);
+      MessageSerializer.intToBytes(offsetBufferCapacity, sizeBytes);
+      out.write(sizeBytes);
 
       out.write(dataBuffer.slice(0, dataBufferCapacity));
       out.write(validityBuffer.slice(0, validityBufferCapacity));
       out.write(offsetBuffer.slice(0, offsetBufferCapacity));
     } else {
-      //TODO support other type
-      throw new UnsupportedOperationException("Unsupported FieldVector:" + vector.getField().getName());
+      throw new UnsupportedOperationException("Only BaseFixedWidthVector and BaseVariableWidthVector are supported.");
     }
   }
 
+  /**
+   * read a vector from channel
+   * @param in the channel to read data
+   * @param vector the vector to read data in
+   * @throws IOException
+   */
   public static void deserializeFieldVector(ReadChannel in, FieldVector vector) throws IOException {
 
-    byte[] reuseIntBytes = new byte[4];
-    in.read(reuseIntBytes);
-    int valueCount = MessageSerializer.bytesToInt(reuseIntBytes);
+    byte[] sizeBytes = new byte[4];
+    in.read(sizeBytes);
+    int valueCount = MessageSerializer.bytesToInt(sizeBytes);
     vector.setValueCount(valueCount);
 
     if (vector instanceof BaseFixedWidthVector) {
-      in.read(reuseIntBytes);
-      int dataLength = MessageSerializer.bytesToInt(reuseIntBytes);
-      in.read(reuseIntBytes);
-      int validityLength = MessageSerializer.bytesToInt(reuseIntBytes);
+      in.read(sizeBytes);
+      int dataLength = MessageSerializer.bytesToInt(sizeBytes);
+      in.read(sizeBytes);
+      int validityLength = MessageSerializer.bytesToInt(sizeBytes);
 
       ArrowBuf dataBuffer = vector.getDataBuffer();
       ArrowBuf validityBuffer = vector.getValidityBuffer();
@@ -95,12 +117,12 @@ public class FieldVectorUtility {
       in.readFully(validityBuffer.nioBuffer(0, validityLength));
 
     } else if (vector instanceof BaseVariableWidthVector) {
-      in.read(reuseIntBytes);
-      int dataLength = MessageSerializer.bytesToInt(reuseIntBytes);
-      in.read(reuseIntBytes);
-      int validityLength = MessageSerializer.bytesToInt(reuseIntBytes);
-      in.read(reuseIntBytes);
-      int offsetLength = MessageSerializer.bytesToInt(reuseIntBytes);
+      in.read(sizeBytes);
+      int dataLength = MessageSerializer.bytesToInt(sizeBytes);
+      in.read(sizeBytes);
+      int validityLength = MessageSerializer.bytesToInt(sizeBytes);
+      in.read(sizeBytes);
+      int offsetLength = MessageSerializer.bytesToInt(sizeBytes);
 
       //resize dataBuffer if needed.
       while (vector.getDataBuffer().capacity() < dataLength) {
@@ -115,7 +137,7 @@ public class FieldVectorUtility {
       in.readFully(validityBuffer.nioBuffer(0, validityLength));
       in.readFully(offsetBuffer.nioBuffer(0, offsetLength));
     } else {
-      throw new UnsupportedOperationException("Unsupported FieldVector:" + vector.getField().getName());
+      throw new UnsupportedOperationException("Only BaseFixedWidthVector and BaseVariableWidthVector are supported.");
     }
   }
 }
